@@ -1,24 +1,7 @@
 {
   "name": "rocksdb",
   "description": "A Node.js RocksDB binding, a backend for LevelUP",
-  "version": "2.0.0",
-  "contributors": [
-    "Rod Vagg <r@va.gg> (https://github.com/rvagg)",
-    "John Chesley <john@chesl.es> (https://github.com/chesles/)",
-    "Jake Verbaten <raynos2@gmail.com> (https://github.com/raynos)",
-    "Dominic Tarr <dominic.tarr@gmail.com> (https://github.com/dominictarr)",
-    "Max Ogden <max@maxogden.com> (https://github.com/maxogden)",
-    "Lars-Magnus Skog <ralphtheninja@riseup.net> (https://github.com/ralphtheninja)",
-    "David Björklund <david.bjorklund@gmail.com> (https://github.com/kesla)",
-    "Julian Gruber <julian@juliangruber.com> (https://github.com/juliangruber)",
-    "Paolo Fragomeni <paolo@async.ly> (https://github.com/hij1nx)",
-    "Anton Whalley <anton.whalley@nearform.com> (https://github.com/No9)",
-    "Matteo Collina <matteo.collina@gmail.com> (https://github.com/mcollina)",
-    "Pedro Teixeira <pedro.teixeira@gmail.com> (https://github.com/pgte)",
-    "James Halliday <mail@substack.net> (https://github.com/substack)",
-    "Gordon Hall <gordonh@member.fsf.org> (https://github.com/bookchin)",
-    "Yichao 'Peak' Ji <peakji@gmail.com> (https://github.com/peakji)"
-  ],
+  "version": "3.0.1",
   "repository": {
     "type": "git",
     "url": "https://github.com/level/rocksdb.git"
@@ -29,42 +12,49 @@
     "level"
   ],
   "main": "leveldown.js",
-  "typings": "leveldown.d.ts",
   "dependencies": {
-    "abstract-leveldown": "~4.0.0",
+    "abstract-leveldown": "~5.0.0",
     "bindings": "~1.3.0",
     "fast-future": "~1.0.2",
-    "nan": "~2.9.1",
-    "prebuild-install": "^2.1.0"
+    "nan": "~2.10.0",
+    "prebuild-install": "^4.0.0"
   },
   "devDependencies": {
     "async": "^2.0.1",
+    "coveralls": "^3.0.2",
     "delayed": "~1.0.1",
     "du": "~0.1.0",
     "iota-array": "~1.0.0",
     "lexicographic-integer": "~1.1.0",
     "mkfiletree": "~1.0.1",
     "monotonic-timestamp": "~0.0.8",
-    "node-uuid": "~1.4.3",
+    "nyc": "^12.0.2",
     "optimist": "~0.6.1",
     "prebuild": "^7.0.0",
     "prebuild-ci": "^2.0.0",
     "readfiletree": "~0.0.1",
     "rimraf": "^2.6.1",
     "slump": "~2.0.0",
-    "tap-spec": "^4.1.1",
+    "standard": "^11.0.1",
     "tape": "^4.5.1",
-    "verify-travis-appveyor": "^2.0.2"
+    "uuid": "^3.2.1",
+    "verify-travis-appveyor": "^3.0.0"
   },
   "scripts": {
     "install": "prebuild-install || node-gyp rebuild",
-    "test": "verify-travis-appveyor && tape test/*-test.js | tap-spec && prebuild-ci",
+    "test": "standard && verify-travis-appveyor && nyc tape test/*-test.js && prebuild-ci",
+    "coverage": "nyc report --reporter=text-lcov | coveralls",
     "rebuild": "prebuild --compile",
     "prebuild": "prebuild -t 8.1.2 -t 6.11.0 -t 4.8.3 -t 7.10.0 --strip --verbose"
   },
   "license": "MIT",
   "gypfile": true,
   "engines": {
-    "node": ">=4"
+    "node": ">=6"
+  },
+  "standard": {
+    "ignore": [
+      "deps/*"
+    ]
   }
 }
