@@ -1,16 +1,18 @@
+/* global gc */
+
 const BUFFERS = false
 
 var leveldown = require('..')
-  , crypto    = require('crypto')
-  , putCount  = 0
-  , getCount  = 0
-  , rssBase
-  , db
+var crypto = require('crypto')
+var putCount = 0
+var getCount = 0
+var rssBase
+var db
 
 function run () {
   var key = 'long key to test memory usage ' + String(Math.floor(Math.random() * 10000000))
 
-  if (BUFFERS) key = new Buffer(key)
+  if (BUFFERS) key = Buffer.from(key)
 
   db.get(key, function (err, value) {
     getCount++
@@ -29,19 +31,13 @@ function run () {
   })
 
   if (getCount % 1000 === 0) {
-    if (typeof gc != 'undefined')
-      gc()
-    console.log(
-        'getCount ='
-      , getCount
-      , ', putCount = '
-      , putCount
-      , ', rss ='
-      , Math.round(process.memoryUsage().rss / rssBase * 100) + '%'
-      , Math.round(process.memoryUsage().rss / 1024 / 1024) + 'M'
-      , JSON.stringify([0,1,2,3,4,5,6].map(function (l) {
-          return db.getProperty('leveldb.num-files-at-level' + l)
-        }))
+    if (typeof gc !== 'undefined') gc()
+    console.log('getCount =', getCount, ', putCount = ', putCount, ', rss =',
+      Math.round(process.memoryUsage().rss / rssBase * 100) + '%',
+      Math.round(process.memoryUsage().rss / 1024 / 1024) + 'M',
+      JSON.stringify([0, 1, 2, 3, 4, 5, 6].map(function (l) {
+        return db.getProperty('leveldb.num-files-at-level' + l)
+      }))
     )
   }
 }
