@@ -1,58 +1,52 @@
-const leveldown   = require('../')
-    , crypto      = require('crypto')
-    , fs          = require('fs')
-    , du          = require('du')
-    , uuid        = require('node-uuid')
+const leveldown = require('../')
+const crypto = require('crypto')
+const fs = require('fs')
+const du = require('du')
+const uuid = require('uuid')
 
-    , entryCount  = 10000000
-    , concurrency = 10
-    , timesFile   = './write_random_times.csv'
-    , dbDir       = './write_random.db'
-    , data        = crypto.randomBytes(256) // buffer
+const entryCount = 10000000
+const concurrency = 10
+const timesFile = './write_random_times.csv'
+const dbDir = './write_random.db'
+const data = crypto.randomBytes(256) // buffer
 
-var db          = leveldown(dbDir)
-  , timesStream = fs.createWriteStream(timesFile, 'utf8')
+var db = leveldown(dbDir)
+var timesStream = fs.createWriteStream(timesFile, 'utf8')
 
 function report (ms) {
   console.log('Wrote', entryCount, 'in', Math.floor(ms / 1000) + 's')
   timesStream.end()
   du(dbDir, function (err, size) {
-    if (err)
-      throw err
+    if (err) throw err
     console.log('Database size:', Math.floor(size / 1024 / 1024) + 'M')
   })
   console.log('Wrote times to ', timesFile)
 }
 
 db.open(function (err) {
-  if (err)
-    throw err
+  if (err) throw err
 
-  var inProgress  = 0
-    , totalWrites = 0
-    , startTime   = Date.now()
-    , writeBuf    = ''
+  var inProgress = 0
+  var totalWrites = 0
+  var startTime = Date.now()
+  var writeBuf = ''
 
-  function write() {
-    if (totalWrites % 100000 == 0) console.log(inProgress, totalWrites)
+  function write () {
+    if (totalWrites % 100000 === 0) console.log(inProgress, totalWrites)
 
-    if (totalWrites % 1000 == 0) {
+    if (totalWrites % 1000 === 0) {
       timesStream.write(writeBuf)
       writeBuf = ''
     }
 
-    if (totalWrites++ == entryCount)
-      return report(Date.now() - startTime)
-
-    if (inProgress >= concurrency || totalWrites > entryCount)
-      return
+    if (totalWrites++ === entryCount) return report(Date.now() - startTime)
+    if (inProgress >= concurrency || totalWrites > entryCount) return
 
     var time = process.hrtime()
     inProgress++
 
     db.put(uuid.v4(), data, function (err) {
-      if (err)
-        throw err
+      if (err) throw err
       writeBuf += (Date.now() - startTime) + ',' + process.hrtime(time)[1] + '\n'
       inProgress--
       process.nextTick(write)
