@@ -1,29 +1,31 @@
-const util             = require('util')
-    , AbstractIterator = require('abstract-leveldown').AbstractIterator
-    , fastFuture       = require('fast-future')
-
+const util = require('util')
+const AbstractIterator = require('abstract-leveldown').AbstractIterator
+const fastFuture = require('fast-future')
 
 function Iterator (db, options) {
   AbstractIterator.call(this, db)
 
-  this.binding    = db.binding.iterator(options)
-  this.cache      = null
-  this.finished   = false
+  this.binding = db.binding.iterator(options)
+  this.cache = null
+  this.finished = false
   this.fastFuture = fastFuture()
 }
 
 util.inherits(Iterator, AbstractIterator)
 
 Iterator.prototype.seek = function (target) {
-  if (this._ended)
+  if (this._ended) {
     throw new Error('cannot call seek() after end()')
-  if (this._nexting)
+  }
+  if (this._nexting) {
     throw new Error('cannot call seek() before next() has completed')
-
-  if (typeof target !== 'string' && !Buffer.isBuffer(target))
+  }
+  if (typeof target !== 'string' && !Buffer.isBuffer(target)) {
     throw new Error('seek() requires a string or buffer key')
-  if (target.length == 0)
+  }
+  if (target.length === 0) {
     throw new Error('cannot seek() to an empty key')
+  }
 
   this.cache = null
   this.binding.seek(target)
@@ -32,17 +34,16 @@ Iterator.prototype.seek = function (target) {
 
 Iterator.prototype._next = function (callback) {
   var that = this
-    , key
-    , value
+  var key
+  var value
 
   if (this.cache && this.cache.length) {
-    key   = this.cache.pop()
+    key = this.cache.pop()
     value = this.cache.pop()
 
     this.fastFuture(function () {
       callback(null, key, value)
     })
-
   } else if (this.finished) {
     this.fastFuture(function () {
       callback()
@@ -51,7 +52,7 @@ Iterator.prototype._next = function (callback) {
     this.binding.next(function (err, array, finished) {
       if (err) return callback(err)
 
-      that.cache    = array
+      that.cache = array
       that.finished = finished
       that._next(callback)
     })
@@ -60,11 +61,9 @@ Iterator.prototype._next = function (callback) {
   return this
 }
 
-
 Iterator.prototype._end = function (callback) {
   delete this.cache
   this.binding.end(callback)
 }
-
 
 module.exports = Iterator
