@@ -1,8 +1,3 @@
-/* Copyright (c) 2012-2017 LevelDOWN contributors
- * See list at <https://github.com/level/leveldown#contributing>
- * MIT License <https://github.com/level/leveldown/blob/master/LICENSE.md>
- */
-
 #include <node.h>
 #include <node_buffer.h>
 
@@ -41,7 +36,7 @@ OpenWorker::OpenWorker (
   , uint32_t maxOpenFiles
   , uint32_t blockRestartInterval
   , uint32_t maxFileSize
-) : AsyncWorker(database, callback)
+) : AsyncWorker(database, callback, "rocksdb:db.open")
 {
   rocksdb::LevelDBOptions levelOptions;
 
@@ -98,8 +93,8 @@ void OpenWorker::Execute () {
 CloseWorker::CloseWorker (
     Database *database
   , Nan::Callback *callback
-) : AsyncWorker(database, callback)
-{};
+) : AsyncWorker(database, callback, "rocksdb:db.close")
+{}
 
 CloseWorker::~CloseWorker () {}
 
@@ -119,9 +114,10 @@ void CloseWorker::WorkComplete () {
 IOWorker::IOWorker (
     Database *database
   , Nan::Callback *callback
+  , const char *resource_name
   , rocksdb::Slice key
   , v8::Local<v8::Object> &keyHandle
-) : AsyncWorker(database, callback)
+) : AsyncWorker(database, callback, resource_name)
   , key(key)
 {
   Nan::HandleScope scope;
@@ -147,7 +143,7 @@ ReadWorker::ReadWorker (
   , bool asBuffer
   , bool fillCache
   , v8::Local<v8::Object> &keyHandle
-) : IOWorker(database, callback, key, keyHandle)
+) : IOWorker(database, callback, "rocksdb:db.get", key, keyHandle)
   , asBuffer(asBuffer)
 {
   Nan::HandleScope scope;
@@ -181,7 +177,7 @@ void ReadWorker::HandleOKCallback () {
       Nan::Null()
     , returnValue
   };
-  callback->Call(2, argv);
+  callback->Call(2, argv, async_resource);
 }
 
 /** DELETE WORKER **/
@@ -192,7 +188,8 @@ DeleteWorker::DeleteWorker (
   , rocksdb::Slice key
   , bool sync
   , v8::Local<v8::Object> &keyHandle
-) : IOWorker(database, callback, key, keyHandle)
+  , const char *resource_name
+) : IOWorker(database, callback, resource_name, key, keyHandle)
 {
   Nan::HandleScope scope;
 
@@ -219,7 +216,7 @@ WriteWorker::WriteWorker (
   , bool sync
   , v8::Local<v8::Object> &keyHandle
   , v8::Local<v8::Object> &valueHandle
-) : DeleteWorker(database, callback, key, sync, keyHandle)
+) : DeleteWorker(database, callback, key, sync, keyHandle, "rocksdb:db.put")
   , value(value)
 {
   Nan::HandleScope scope;
@@ -247,7 +244,7 @@ BatchWorker::BatchWorker (
   , Nan::Callback *callback
   , rocksdb::WriteBatch* batch
   , bool sync
-) : AsyncWorker(database, callback)
+) : AsyncWorker(database, callback, "rocksdb:db.batch")
   , batch(batch)
 {
   options = new rocksdb::WriteOptions();
@@ -272,7 +269,7 @@ ApproximateSizeWorker::ApproximateSizeWorker (
   , rocksdb::Slice end
   , v8::Local<v8::Object> &startHandle
   , v8::Local<v8::Object> &endHandle
-) : AsyncWorker(database, callback)
+) : AsyncWorker(database, callback, "rocksdb:db.approximateSize")
   , range(start, end)
 {
   Nan::HandleScope scope;
@@ -303,7 +300,7 @@ void ApproximateSizeWorker::HandleOKCallback () {
       Nan::Null()
     , returnValue
   };
-  callback->Call(2, argv);
+  callback->Call(2, argv, async_resource);
 }
 
 /** COMPACT RANGE WORKER **/
@@ -315,7 +312,7 @@ CompactRangeWorker::CompactRangeWorker (
   , rocksdb::Slice end
   , v8::Local<v8::Object> &startHandle
   , v8::Local<v8::Object> &endHandle
-) : AsyncWorker(database, callback)
+) : AsyncWorker(database, callback, "rocksdb:db.compactRange")
 {
   Nan::HandleScope scope;
 
@@ -346,7 +343,7 @@ void CompactRangeWorker::HandleOKCallback () {
   v8::Local<v8::Value> argv[] = {
       Nan::Null()
   };
-  callback->Call(1, argv);
+  callback->Call(1, argv, async_resource);
 }
 
 } // namespace leveldown
