@@ -1,7 +1,7 @@
-const test       = require('tape')
-    , testCommon = require('abstract-leveldown/testCommon')
-    , leveldown  = require('../')
-    , abstract   = require('abstract-leveldown/abstract/close-test')
+const test = require('tape')
+const testCommon = require('abstract-leveldown/testCommon')
+const leveldown = require('../')
+const abstract = require('abstract-leveldown/abstract/close-test')
 
 module.exports.setUp = function () {
   test('setUp', testCommon.setUp)
