@@ -1,6 +1,6 @@
-const test       = require('tape')
-    , testCommon = require('abstract-leveldown/testCommon')
-    , leveldown  = require('../')
+const test = require('tape')
+const testCommon = require('abstract-leveldown/testCommon')
+const leveldown = require('../')
 
 var db
 
@@ -12,20 +12,18 @@ test('setUp db', function (t) {
 })
 
 test('test argument-less getProperty() throws', function (t) {
-  t.throws(
-      db.getProperty.bind(db)
-    , { name: 'Error', message: 'getProperty() requires a valid `property` argument' }
-    , 'no-arg getProperty() throws'
-  )
+  t.throws(db.getProperty.bind(db), {
+    name: 'Error',
+    message: 'getProperty() requires a valid `property` argument'
+  }, 'no-arg getProperty() throws')
   t.end()
 })
 
 test('test non-string getProperty() throws', function (t) {
-  t.throws(
-      db.getProperty.bind(db, {})
-    , { name: 'Error', message: 'getProperty() requires a valid `property` argument' }
-    , 'no-arg getProperty() throws'
-  )
+  t.throws(db.getProperty.bind(db, {}), {
+    name: 'Error',
+    message: 'getProperty() requires a valid `property` argument'
+  }, 'no-arg getProperty() throws')
   t.end()
 })
 
@@ -36,19 +34,21 @@ test('test invalid getProperty() returns empty string', function (t) {
 })
 
 test('test invalid getProperty("rocksdb.num-files-at-levelN") returns numbers', function (t) {
-  for (var i = 0; i < 7; i++)
+  for (var i = 0; i < 7; i++) {
     t.equal(db.getProperty('rocksdb.num-files-at-level' + i), '0', '"rocksdb.num-files-at-levelN" === "0"')
+  }
   t.end()
 })
 
 test('test invalid getProperty("rocksdb.stats")', function (t) {
-  debugger;
   t.ok(db.getProperty('rocksdb.stats').split('\n').length > 3, 'rocksdb.stats has > 3 newlines')
   t.end()
 })
 
 test('test invalid getProperty("rocksdb.sstables")', function (t) {
-  var expected = [0,1,2,3,4,5,6].map(function (l) { return '--- level ' + l + ' --- version# 1 ---' }).join('\n') + '\n'
+  var expected = [0, 1, 2, 3, 4, 5, 6].map(function (l) {
+    return '--- level ' + l + ' --- version# 1 ---'
+  }).join('\n') + '\n'
   t.equal(db.getProperty('rocksdb.sstables'), expected, 'rocksdb.sstables')
   t.end()
 })
