@@ -1,5 +1,5 @@
-const test       = require('tape')
-    , leveldown  = require('..')
-    , abstract   = require('abstract-leveldown/abstract/iterator-range-test')
+const test = require('tape')
+const leveldown = require('..')
+const abstract = require('abstract-leveldown/abstract/iterator-range-test')
 
 abstract.all(leveldown, test)
