@@ -1,5 +1,5 @@
-const test      = require('tape')
-    , leveldown = require('../')
-    , abstract  = require('abstract-leveldown/abstract/leveldown-test')
+const test = require('tape')
+const leveldown = require('../')
+const abstract = require('abstract-leveldown/abstract/leveldown-test')
 
 abstract.args(leveldown, test)
