@@ -1,14 +1,14 @@
-const test       = require('tape')
-    , testCommon = require('abstract-leveldown/testCommon')
-    , cleanup    = testCommon.cleanup
-    , location   = testCommon.location
-    , leveldown  = require('../')
+const test = require('tape')
+const testCommon = require('abstract-leveldown/testCommon')
+const cleanup = testCommon.cleanup
+const location = testCommon.location
+const leveldown = require('../')
 
 function makeTest (name, testFn) {
   test(name, function (t) {
     cleanup(function () {
-      var loc  = location()
-      var db   = leveldown(loc)
+      var loc = location()
+      var db = leveldown(loc)
       var done = function (close) {
         if (close === false) {
           cleanup(function (err) {
