@@ -1,5 +1,5 @@
-const test       = require('tape')
-    , leveldown  = require('../')
-    , abstract   = require('abstract-leveldown/abstract/open-test')
+const test = require('tape')
+const leveldown = require('../')
+const abstract = require('abstract-leveldown/abstract/open-test')
 
 abstract.all(leveldown, test)
