@@ -1,26 +1,21 @@
-const test         = require('tape')
-    , fs           = require('fs')
-    , path         = require('path')
-    , mkfiletree   = require('mkfiletree')
-    , readfiletree = require('readfiletree')
-    , leveldown    = require('../')
-    , makeTest     = require('./make')
+const test = require('tape')
+const fs = require('fs')
+const leveldown = require('../')
+const makeTest = require('./make')
 
 test('test argument-less repair() throws', function (t) {
-  t.throws(
-      leveldown.repair
-    , { name: 'Error', message: 'repair() requires `location` and `callback` arguments' }
-    , 'no-arg repair() throws'
-  )
+  t.throws(leveldown.repair, {
+    name: 'Error',
+    message: 'repair() requires `location` and `callback` arguments'
+  }, 'no-arg repair() throws')
   t.end()
 })
 
 test('test callback-less, 1-arg, repair() throws', function (t) {
-  t.throws(
-      leveldown.repair.bind(null, 'foo')
-    , { name: 'Error', message: 'repair() requires `location` and `callback` arguments' }
-    , 'callback-less, 1-arg repair() throws'
-  )
+  t.throws(leveldown.repair.bind(null, 'foo'), {
+    name: 'Error',
+    message: 'repair() requires `location` and `callback` arguments'
+  }, 'callback-less, 1-arg repair() throws')
   t.end()
 })
 
